# docs/api-server/src/lib/api-zod/generated/types/calendarSelection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Outlook to Google Calendar Sync API
 * OpenAPI spec version: 1.1.0
 */

export interface CalendarSelection {
  calendarId: string;
}