# docs/api-server/src/lib/api-zod/generated/types/outlookConnection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Outlook to Google Calendar Sync API
 * OpenAPI spec version: 1.1.0
 */

export interface OutlookConnection {
  profilePath: string;
}