# docs/api-server/src/lib/api-zod/generated/types/syncRun.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Outlook to Google Calendar Sync API
 * OpenAPI spec version: 1.1.0
 */
import type { SyncRunStatus } from "./syncRunStatus";

export interface SyncRun {
  id: string;
  startedAt: Date;
  finishedAt?: Date;
  status: SyncRunStatus;
  created?: number;
  updated?: number;
  deleted?: number;
  message?: string;
}