# docs/api-server/src/lib/api-zod/generated/types/syncRunDetails.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Outlook to Google Calendar Sync API
 * OpenAPI spec version: 1.1.0
 */
import type { LogEntry } from "./logEntry";
import type { SyncRun } from "./syncRun";

export interface SyncRunDetails {
  run: SyncRun;
  entries: LogEntry[];
}