# docs/calendar-sync/src/lib/api-client-react/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Outlook to Google Calendar Sync API
 * OpenAPI spec version: 1.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  googleConnected: boolean;
  outlookConnected: boolean;
  outlookProfilePath?: string;
  outlookCalendarId?: string;
  outlookCalendarName?: string;
  googleCalendarId?: string;
  googleCalendarName?: string;
}

export interface OutlookConnection {
  profilePath: string;
}

export interface CalendarRef {
  id: string;
  name: string;
  timeZone?: string;
  color?: string;
}

export interface CalendarSelection {
  calendarId: string;
}

export interface SyncSettings {
  /** @minimum 1 */
  frequencyMinutes: number;
  /** @minimum 0 */
  daysPast: number;
  /** @minimum 0 */
  daysFuture: number;
  debugLogging: boolean;
}

export type SyncRunStatus = (typeof SyncRunStatus)[keyof typeof SyncRunStatus];

export const SyncRunStatus = {
  SUCCESS: "SUCCESS",
  FAILED: "FAILED",
} as const;

export interface SyncRun {
  id: string;
  startedAt: string;
  finishedAt?: string;
  status: SyncRunStatus;
  created?: number;
  updated?: number;
  deleted?: number;
  message?: string;
}

export interface PagedSyncRuns {
  content: SyncRun[];
  page: number;
  size: number;
  totalElements: number;
  totalPages: number;
}

export type LogEntryLevel = (typeof LogEntryLevel)[keyof typeof LogEntryLevel];

export const LogEntryLevel = {
  INFO: "INFO",
  DEBUG: "DEBUG",
  WARN: "WARN",
  ERROR: "ERROR",
} as const;

export interface LogEntry {
  timestamp: string;
  level: LogEntryLevel;
  message: string;
}

export interface SyncRunDetails {
  run: SyncRun;
  entries: LogEntry[];
}

export type GetLogsParams = {
  /**
   * @minimum 0
   */
  page?: number;
  /**
   * @minimum 1
   */
  size?: number;
  status?: GetLogsStatus;
};

export type GetLogsStatus = (typeof GetLogsStatus)[keyof typeof GetLogsStatus];

export const GetLogsStatus = {
  SUCCESS: "SUCCESS",
  FAILED: "FAILED",
} as const;